fix: update IEmailService.SendJobExecutionNotificationAsync to accept long jobExecutionId
- Changed parameter type from int to long to match JobExecution.Id type
- Updated both interface and implementation
- Fixes CS1503 compilation errors in job classes

+++ src/SchedulerPlatform.Core/Domain/Interfaces/IEmailService.cs
@@ -4,5 +4,5 @@ public interface IEmailService
 {
     Task SendEmailAsync(string to, string subject, string body, bool isHtml = true);
     Task SendEmailAsync(IEnumerable<string> recipients, string subject, string body, bool isHtml = true);
-    Task SendJobExecutionNotificationAsync(int jobExecutionId, bool isSuccess);
+    Task SendJobExecutionNotificationAsync(long jobExecutionId, bool isSuccess);
 }
